замена сервиса  обработки сообщений
